🧹 Jules05: Technical debt cleanup — architectural harmonization
Consolidated fragmented risk management logic by porting the 8-layer
safety cascade and ATR-based sizing from RiskEngine to RiskManager.
Decommissioned redundant src/trading/risk_engine.py and updated
main.py and AuditedRiskManager to use the harmonized validate_signal API.

Remediated technical debt by removing stale test files and addressing
multi-agent drift. Verified integrity with the harmonized test suite
and core product coherence checks.

Key changes:
- Harmonized RiskManager with 8-layer cascade.
- Updated main.py to pass market context and register PnL.
- Removed 6 redundant/stale test files.
- Documented findings in docs/technical-debt/DEBT_LOG.md.
